Rename UI translation files
