+++ src/i18n/ar/ui.ts
