+++ src/i18n/bg/ui.ts
