+++ src/i18n/bn/ui.ts
