+++ src/i18n/da/ui.ts
