+++ src/i18n/es/ui.ts
