+++ src/i18n/fr/ui.ts
