+++ src/i18n/ja/ui.ts
