+++ src/i18n/kr/ui.ts
