+++ src/i18n/zh-CN/ui.ts
